fix: padronizar size_bytes como int64 e alinhar mensagens de PDF e binario
size_bytes divergia entre os caminhos de read_file; aviso de PDF prometia document_mode ocr inexistente na Fase 1; recusa de binario grande reaproveita a mensagem da extracao.

+++ internal/tools/filesystem/read_file.go
@@ -125,7 +125,7 @@ func (t *ReadFile) Execute(ctx context.Context, args json.RawMessage) (tools.Too
 		meta = map[string]any{
 			"projection": true,
 			"format":     string(extracted.Kind),
-			"size_bytes": len(data),
+			"size_bytes": int64(len(data)),
 		}
 		if extracted.Pages > 0 {
 			meta["pages"] = extracted.Pages
@@ -133,7 +133,7 @@ func (t *ReadFile) Execute(ctx context.Context, args json.RawMessage) (tools.Too
 	} else {
 		content = extracted.Markdown
 		meta = map[string]any{
-			"size_bytes": len(data),
+			"size_bytes": int64(len(data)),
 		}
 	}
